Добавить бд-репозитории, экстракторы, поменять сервис и контроллеры для работы с БД, добавить новые поля в User и Film. Добавить Mpa и Genre.

+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -15,7 +15,7 @@ import java.util.Set;
 @NoArgsConstructor
 public class User {
 
-	private Integer id;
+	private Long id;
 	@NotEmpty(message = "E-mail не должен быть пуст", groups = BasicInfo.class)
 	@Email(message = "Неправильный формат e-mail.", groups = AdvanceInfo.class)
 	private String email;
@@ -26,7 +26,7 @@ public class User {
 	@Past(groups = AdvanceInfo.class)
 	@JsonFormat(pattern = "yyyy-MM-dd")
 	private LocalDate birthday;
-	private Set<Integer> friends = new HashSet<>();
-	private Set<Integer> likedFilms = new HashSet<>();
+	private Set<Long> friends = new HashSet<>();
+	private Set<Long> likedFilms = new HashSet<>();
 
 }
